modulation of written code
